💄 refactor(chat): dock selection toolbar at the bottom
Restore the selection action bar to a full-width bar docked at the bottom in
place of the composer, instead of the floating pill. Keeps the count + Cancel /
Delete / Forward actions, forward dialog, and Esc-to-exit behavior.
