+++ src/features/Conversation/MessageForward/MessageForwardFooter.tsx
@@ -3,17 +3,16 @@
 import { memo, type ReactNode } from 'react';
 
 import { messageStateSelectors, useConversationStore } from '../store';
-import SelectionFloatBar from './SelectionFloatBar';
+import SelectionFooterBar from './SelectionFooterBar';
 
 interface MessageForwardFooterProps {
   children: ReactNode;
 }
 
 /**
- * Wraps the chat composer: while multi-selecting it hides the input and floats
- * the selection action pill over the bottom of the conversation. The input stays
- * mounted but display:none so the editor/draft state survives toggling selection
- * mode.
+ * Wraps the chat composer: while multi-selecting it hides the input and docks
+ * the selection action bar at the bottom in its place. The input stays mounted
+ * but display:none so the editor/draft state survives toggling selection mode.
  */
 const MessageForwardFooter = memo<MessageForwardFooterProps>(({ children }) => {
   const isSelectionMode = useConversationStore(messageStateSelectors.isSelectionMode);
@@ -21,7 +20,7 @@ const MessageForwardFooter = memo<MessageForwardFooterProps>(({ children }) => {
   return (
     <>
       <div style={isSelectionMode ? { display: 'none' } : { display: 'contents' }}>{children}</div>
-      {isSelectionMode && <SelectionFloatBar />}
+      {isSelectionMode && <SelectionFooterBar />}
     </>
   );
 });
